arbion-backend/django-engine/requirements.txt
Django
gunicorn
psycopg2-binary
django-cors-headers
python-dotenv
djangorestframework
djangorestframework-simplejwt
whitenoise
psycopg2-binary
request
